backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
supabase>=2,<3
slowapi>=0.1.9,<1
pydantic-settings>=2,<3
python-dotenv>=1,<2
